+++ src/app/styles/admin/eventPreview.module.css
@@ -1,29 +1,59 @@
 .eventCard {
-  background-color: #fff; /* Light background */
-  border-radius: 8px; /* Rounded corners */
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1); /* Subtle shadow for depth */
-  padding: 20px; /* Spacing inside the card */
-  margin-bottom: 20px; /* Space between cards when listed */
-  transition: transform 0.3s ease; /* Smooth transition for hover effect */
+  background-color: #e7dabf;
+  border-radius: 12px;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  padding: 20px;
+  transition: transform 0.3s ease;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-around;
+  align-items: center;
+  width: 30vw;
+  height: 250px;
+}
+
+.eventCard h2 {
+  color: #2b2b34;
+  font-size: 2vh;
+  font-weight: bold;
 }
 
 .eventCard h3 {
-  color: #333; 
-  margin-top: 0; 
+  color: #2b2b34;
+  font-weight: 275;
+  font-size: 1.2vh;
 }
 
-.eventCard p {
-  color: #666; 
-  line-height: 1.5; 
-  margin: 10px 0;
-}
-
-.eventCard p strong {
-  color: #333; /* Make labels darker for contrast */
-}
-
-/* Hover effect to lift the card */
 .eventCard:hover {
   transform: translateY(-5px);
   box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
+  cursor: pointer;
+}
+
+.imageContainer {
+  display: flex;
+  width: 50%;
+  justify-content: center;
+}
+
+.imageWrapper {
+  border-radius: 50%;
+  width: 150px;
+  height: 150px;
+  position: relative;
+  overflow: hidden;
+}
+
+.infoContainer {
+  width: 50%;
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  justify-content: space-between;
+}
+
+@media (max-width: 768px) {
+  .eventCard {
+    width: 65vw;
+  }
 }
